+++ src/jOpqua.jl
@@ -7,6 +7,13 @@ include("utils.jl")
 include("simulation/constants.jl")
 include("simulation/structs.jl")
 
+# Extensions
+include("extensions/immunity/immunity.jl")
+include("extensions/intrahost/intrahost.jl")
+
+# Input
+include("parameters/parameters.jl")
+
 # Simulation
 include("simulation/initializers.jl")
 include("simulation/flexle.jl")
@@ -15,15 +22,8 @@ include("simulation/weights.jl")
 include("simulation/events.jl")
 include("simulation/simulation.jl")
 
-# Input
-include("parameters/parameters.jl")
-
 # Output
 include("analysis/data.jl")
 include("analysis/plots.jl")
-
-# Extensions
-include("extensions/immunity/immunity.jl")
-include("extensions/intrahost/intrahost.jl")
 
 end # module jOpqua