+++ src/parameters/parameters.jl
@@ -1,57 +1,55 @@
 using StaticArrays
 
-function newPathogenType(id::String;)
-    return PathogenType(
-        id,
-        10,
-        "ARNDCEQGHILKMFPSTWYV*",
-        1.0,
-        0.0,
-        0.0,
-        0.0,
-        SA[ # pathogen_coefficient_functions::SVector{NUM_COEFFICIENTS,Function} # Each element takes seq argument, returns Float64
-            g->1.0, g->1.0, g->1.0, g->1.0,
-            g->1.0, g->1.0, g->1.0, g->1.0,
-            g->1.0, g->1.0, g->1.0, g->1.0,
-        ],
-        g -> 1.0, g -> 1.0, g -> 1.0, g -> 1.0,
-    )
-end
+const DEFAULT_PATHOGEN_TYPE = PathogenType(
+    "Default",
+    10,
+    "ARNDCEQGHILKMFPSTWYV*",
+    1.0,
+    0.0,
+    0.0,
+    0.0,
+    g->1.0, g->1.0, g->1.0, g->1.0,
+    SA[ # order defined in COEFFICIENTS
+        g->1.0, g->1.0, g->1.0,
+        g->1.0, g->1.0, g->1.0,
+        g->1.0, g->1.0, g->1.0,
+        g->1.0, g->1.0, g->1.0
+    ],
+)
 
-function newResponseType(id::String;)
-    return ResponseType(
-        id, # id::String,
-        SA[ # static_coefficient_functions::SVector{NUM_COEFFICIENTS,Function},
-            (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
-            (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
-            (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0, (imp_g, mat_g)->1.0,
-        ],
-        SA[ # specific_coefficient_functions::SVector{NUM_COEFFICIENTS,Function},
-            (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
-            (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
-            (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0, (imp_g, mat_g, pat_g)->1.0,
-        ],
-        0.0,
-        (imp_g, mat_g, pat_g) -> 1.0,
-        (imp_g, mat_g, pat_g) -> 1.0,
-    )
-end
+const DEFAULT_RESPONSE_TYPE = ResponseType(
+    "Default",
+    0.0,
+    (imp_g, mat_g, pat_g) -> 1.0,
+    (imp_g, mat_g, pat_g) -> 1.0,
+    SA[ # order defined in COEFFICIENTS
+        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
+        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
+        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
+        (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0, (imp_g, mat_g) -> 1.0,
+    ],
+    SA[ # order defined in COEFFICIENTS
+        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
+        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
+        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0,
+        (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0, (imp_g, mat_g, pat_g) -> 1.0
+    ],
+)
 
-function newPopulationType(id::String;)
-    return PopulationType(
-        id, # id::String
-        SA[ # base_coefficients::SVector{NUM_COEFFICIENTS,Float64}
-            0.00, 1.0, 0.0, 0.0,
-            1.05, 0.0, 0.0, 0.0,
-            0.00, 0.0, 1.0, 0.0,
-        ],
-        true, false,
-        pathogenFractionsWinnerTakesAll,
-        weightedResponseArithmeticMean,
-        infectionProbabilityArithmeticMean,
-        (p, h, c) -> Nothing, # developResponses::Function,
-        1.0, # inoculum_coefficient
-        1.0, # mutation_coefficient
-        1.0, # recombination_coefficient
-    )
-end
+const DEFAULT_POPULATION_TYPE = PopulationType(
+    "Default", # id::String
+    true, false,
+    1.0, # inoculum_coefficient
+    1.0, # mutation_coefficient
+    1.0, # recombination_coefficient
+    SA[ # base_coefficients::SVector{NUM_COEFFICIENTS,Float64}
+        0.0, 1.0, 0.0,
+        0.0, 1.05, 0.0,
+        0.0, 0.0, 0.0,
+        0.0, 1.0, 0.0
+    ],
+    pathogenFractionsWinnerTakesAll,
+    weightedResponseArithmeticMean,
+    infectionProbabilityArithmeticMean,
+    (p, h, c) -> Nothing, # developResponses::Function,
+)